+++ package/signal/src/signal.ts
@@ -1,6 +1,17 @@
-import {log, _getSignalObject, _callListeners, _removeSignalListener} from './core';
-import type {ListenerOptions, DispatchOptions, ListenerCallback, ListenerObject} from './type';
+import {log,
+  _getSignalObject,
+  _callListeners,
+  _removeSignalListener,
+} from './core';
 
+import type {
+  ListenerOptions,
+  DispatchOptions,
+  ListenerCallback,
+  ListenerObject,
+  SignalProvider,
+  SignalProviderOptions,
+} from './type';
 
 /**
  * Add new listener to specific signal.
@@ -11,13 +22,13 @@ import type {ListenerOptions, DispatchOptions, ListenerCallback, ListenerObject}
 export function addSignalListener<SignalName extends keyof VatrSignals>(
     signalName: SignalName,
     signalCallback: ListenerCallback<SignalName>,
-    options?: Partial<ListenerOptions>,
+    options?: ListenerOptions,
 ): symbol {
   log('addSignalListener(%s, %o)', signalName, options);
 
   const signal = _getSignalObject(signalName);
   const listener: ListenerObject<SignalName> = {
-    id: Symbol('Vatr Signal Listener'),
+    id: Symbol('Vatr signal listener for ' + signalName),
     once: options?.once ?? false,
     disabled: options?.disabled ?? false,
     callback: signalCallback,
@@ -81,7 +92,7 @@ export function removeSignalListener<SignalName extends keyof VatrSignals>(
 export function dispatchSignal<SignalName extends keyof VatrSignals>(
     signalName: SignalName,
     value: VatrSignals[SignalName],
-    options?: Partial<DispatchOptions>,
+    options?: DispatchOptions,
 ): void {
   log('dispatchSignal(%s, %o, %o)', signalName, value, options);
 
@@ -139,17 +150,35 @@ export function requestSignal<SignalName extends keyof VatrRequestSignals>(
  *   }
  * }
  */
-export function addSignalProvider<SignalName extends keyof VatrRequestSignals>(
+export function setSignalProvider<SignalName extends keyof VatrRequestSignals>(
     signalName: SignalName,
-    signalCallback: (detail: VatrRequestSignals[SignalName]) => void | Promise<void>,
+    signalProvider: SignalProvider<SignalName>,
+    options?: SignalProviderOptions,
 ): symbol {
-  log('addSignalProvider(%s)', signalName);
+  log('setSignalProvider(%s)', signalName);
+  // @TODO: refactor with removeSignalProvider
+  const signal = _getSignalObject(`request-${signalName}` as unknown as SignalName);
+  if (signal.listenerList.length > 0) {
+    log('setSignalProvider(%s): WARNING! another provider defined and will removed!',
+        signalName, signal.listenerList.length);
+    signal.listenerList = [];
+  }
+
+  const _callback = async (requestParam: VatrRequestSignals[SignalName]): Promise<void> => {
+    const signalValue = await signalProvider(requestParam);
+    if (signalValue !== undefined) { // null can be a valid value.
+      dispatchSignal(signalName, signalValue, {debounce: options?.debounce ?? true});
+    }
+  };
+
   return addSignalListener(
     `request-${signalName}` as unknown as SignalName,
-    signalCallback as unknown as ListenerCallback<SignalName>,
-    {receivePrevious: true},
+    _callback as unknown as ListenerCallback<SignalName>,
+    {receivePrevious: options?.receivePrevious ?? true},
   );
 }
+
+// @TODO: removeSignalProvider(signalName): void
 
 /**
  * Resolved with signal value when signal is ready base on requested options.